requirements.txt
Pillow==9.5.0
Django==4.1.3
psycopg2-binary==2.9.5
celery==5.2.7
django-celery-beat==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1
drf-nested-routers==0.93.4
requests==2.28.2
beautifulsoup4==4.11.1
lxml==4.9.1
fake-useragent==1.1.1
gunicorn==20.1.0
redis==4.5.1